#17110 Move service test utils to workbench
