+++ .vscode/launch.json
@@ -10,7 +10,7 @@
 			"stopOnEntry": false,
 			"args": [
 				"--timeout",
-				"999999"
+				"3999"
 			],
 			"cwd": "${workspaceRoot}",
 			"env": {
