+++ src/vs/workbench/test/common/editor/resourceEditorInput.test.ts
@@ -10,7 +10,7 @@ import URI from 'vs/base/common/uri';
 import { ResourceEditorInput } from 'vs/workbench/common/editor/resourceEditorInput';
 import { ResourceEditorModel } from 'vs/workbench/common/editor/resourceEditorModel';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { workbenchInstantiationService } from 'vs/test/utils/servicesTestUtils';
+import { workbenchInstantiationService } from 'vs/workbench/test/workbenchTestUtils';
 import { IModelService } from 'vs/editor/common/services/modelService';
 import { IModeService } from 'vs/editor/common/services/modeService';
