+++ src/vs/workbench/test/common/editor/untitledEditor.test.ts
@@ -10,7 +10,7 @@ import { TPromise } from 'vs/base/common/winjs.base';
 import { join } from 'vs/base/common/paths';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IUntitledEditorService, UntitledEditorService } from 'vs/workbench/services/untitled/common/untitledEditorService';
-import { workbenchInstantiationService } from 'vs/test/utils/servicesTestUtils';
+import { workbenchInstantiationService } from 'vs/workbench/test/workbenchTestUtils';
 import { UntitledEditorModel } from 'vs/workbench/common/editor/untitledEditorModel';
 
 class ServiceAccessor {
