+++ src/vs/workbench/test/common/memento.test.ts
@@ -8,7 +8,7 @@
 import * as assert from 'assert';
 import { WorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { StorageScope } from 'vs/platform/storage/common/storage';
-import * as TestUtils from 'vs/test/utils/servicesTestUtils';
+import * as TestUtils from 'vs/workbench/test/workbenchTestUtils';
 import { Memento, Scope } from 'vs/workbench/common/memento';
 import { StorageService, InMemoryLocalStorage } from 'vs/platform/storage/common/storageService';
